taeho_rca.sh: Reduce number of duplicate BMC Info 'records'
Attempting to reduce number of duplicate 'BMC Info' records.

The logs can contain many 10s or 100s of duplicate records. So
attempt to reduce number echo'ed to the screen

Note: bash function is same as that used for FRU - and seems to
be overly complicated for what it is trying to acheive..

Signed-off-by: Mark Harvey <mark.harvey@nutanix.com>